+++ src/pages/Prof/Prof.tsx
@@ -2,17 +2,16 @@ import React, {useEffect, useRef, useState} from "react";
 import FilterForm from "../../component/Filter/filterForm";
 import {ProfFilterForm, profTab} from "./profConsts";
 import TabForm, {btnInetrface} from "../../component/Tableau/tableauxForm";
-import {deleteProf, getProf, getProfWithStatus} from "../../store/modules/Prof/profService";
+import {deleteProf, getProfWithStatus} from "../../store/modules/Prof/profService";
 import {useSelector} from "react-redux";
 import {RootState, store} from "../../store/store";
 import {useNavigate} from "react-router-dom";
-import {setSelectedProf} from "../../store/modules/Prof/profModule";
 import {setUserToHistory} from "../../store/modules/Auth/authService";
 import moment from "moment";
 import {GetDirector, updateDirectorWithMail} from "../../store/modules/Director/directorService";
 import ModalComp from "../../component/Modal/modalComp";
 import {setSelectedDirector} from "../../store/modules/Director/directorModule";
-import {DirectorTab} from "../Director/directorConsts";
+import Swal from "sweetalert2";
 
 export interface detailProf {
 	name: string,
@@ -50,22 +49,60 @@ const Prof = () => {
 	const receive = (data: { index: number, btn: btnInetrface }) => {
 		switch (data.btn?.type) {
 			case 'detail':
-				// store.dispatch(setSelectedProf(listProf[data.index]))
-				// navigate('/prof/detail/:id')
-				// break;
 				store.dispatch(setSelectedDirector(profTab.data[data.index]))
 				setModal(true)
 				break;
 			case 'delete':
-				deleteProf({_id: listProf[data.index]._id}).then(() => setUserToHistory({
-					date: moment().format('MMMM Do YYYY, h:mm:ss a'),
-					adminID: user.user._id,
-					userId: listProf[data.index]._id,
-					data: listProf[data.index],
-					type: 'delete'
-				}).then(() => {
-					getProfWithStatus().then()
-				}))
+				const swalWithBootstrapButtons = Swal.mixin({
+					customClass: {
+						confirmButton: 'btn btn-success',
+						cancelButton: 'btn btn-error'
+					},
+					buttonsStyling: false
+				})
+
+				swalWithBootstrapButtons.fire({
+					title: 'هل انت متأكد؟',
+					text: "هل تريد حظر المستخدم ",
+					icon: 'warning',
+					showCancelButton: true,
+					confirmButtonText: 'نعم',
+					cancelButtonText: 'لا',
+					reverseButtons: true
+				}).then((result) => {
+					if (result.isConfirmed) {
+						deleteProf({_id: listProf[data.index]._id}).then(() => setUserToHistory({
+							date: moment().format('MMMM Do YYYY, h:mm:ss a'),
+							adminID: user.user._id,
+							userId: listProf[data.index]._id,
+							data: listProf[data.index],
+							type: 'delete'
+						}).then(() => {
+							getProfWithStatus().then()
+						}))
+						Swal.fire({
+							position: 'center',
+							icon: 'success',
+							title: 'لقد تم الحظر',
+							showConfirmButton: false,
+							timer: 1500
+						})
+
+					} else if (
+						/* Read more about handling dismissals below */
+						result.dismiss === Swal.DismissReason.cancel
+					) {
+						Swal.fire({
+							position: 'center',
+							icon: 'success',
+							title: 'تم الغاء الحظر',
+							showConfirmButton: false,
+							timer: 1500
+						})
+					}
+				})
+
+
 				break;
 
 		}
@@ -86,13 +123,59 @@ const Prof = () => {
 	const PhoneRegex=/^(009665|9665|\+9665|05|5)(5|0|3|6|4|9|1|8|7)([0-9]{7})$/
 	const onSubmit = (data: { [key: string]: string | number }) => {
 		if (!data.name  ||!data.tel){
-			alert('يجب عليك كاتبة كل الخانات')
-		}else if(!PhoneRegex.test(data.tel.toString())) {
-			alert('الرجاء التحقق من البريد الالكرتروني او الهاتف الجوال')
-		}else{
-			updateDirectorWithMail({...data,mail:selected.mail}).then((res: any) => {
-				getProfWithStatus().then(() => setModal(false))
+			Swal.fire({
+				icon: 'error',
+				title: 'المعلومات غير كافية',
+				text: 'يجب عليك تعمير كل المعلومات',
+			})			}else if(!PhoneRegex.test(data.tel.toString())) {
+			Swal.fire({
+				icon: 'error',
+				title: 'يجب ان يكون رقم من السعودية',
+				text: 'رقم الهاتف الجوال خاطئ',
+			})			}else{
+			const swalWithBootstrapButtons = Swal.mixin({
+				customClass: {
+					confirmButton: 'btn btn-success',
+					cancelButton: 'btn btn-error'
+				},
+				buttonsStyling: false
 			})
+
+			swalWithBootstrapButtons.fire({
+				title: 'هل انت متأكد؟',
+				text: "هل تريد تغير معلومات الأستاذ ",
+				icon: 'warning',
+				showCancelButton: true,
+				confirmButtonText: 'نعم',
+				cancelButtonText: 'لا',
+				reverseButtons: true
+			}).then((result) => {
+				if (result.isConfirmed) {
+					updateDirectorWithMail({...data,mail:selected.mail}).then((res: any) => {
+						GetDirector().then(() => setModal(false))
+					})
+					Swal.fire({
+						position: 'top-end',
+						icon: 'success',
+						title: 'لقد تم تغير المعلومات ',
+						showConfirmButton: false,
+						timer: 1000
+					})
+
+				} else if (
+					/* Read more about handling dismissals below */
+					result.dismiss === Swal.DismissReason.cancel
+				) {
+					Swal.fire({
+						position: 'top-end',
+						icon: 'error',
+						title: 'تم الغاء التغير',
+						showConfirmButton: false,
+						timer: 1000
+					})
+				}
+			})
+
 		}
 
 	}
@@ -125,6 +208,7 @@ const Prof = () => {
 			name: item.name,
 			mail: item.mail,
 			tel: item.tel,
+			spec:item?.specialite
 		}))
 		setTableModel({...temp})
 	}
