+++ src/pages/sigql/signal.tsx
@@ -16,6 +16,7 @@ import {signalTabConst} from "./siganConst";
 import {deleteProf, getProfWithStatus} from "../../store/modules/Prof/profService";
 import {setUserToHistory} from "../../store/modules/Auth/authService";
 import moment from "moment";
+import Swal from "sweetalert2";
 
 const Signal=()=>{
 	const [tableModel, setTableModel] = useState<any>(signalTabConst)
@@ -39,19 +40,58 @@ const Signal=()=>{
 		switch (data.btn?.type) {
 
 			case 'detail':
-				blockUser.current=userId[data.index]
-				deleteProf({_id: blockUser.current}).then(() => setUserToHistory({
-					date: moment().format('MMMM Do YYYY, h:mm:ss a'),
-					adminID: userConnect.user._id,
-					userId: blockUser.current,
-					data: 'شكوى',
-					type: 'delete'
-				})).then(()=>{
-					deleteSignal({id:test[data.index]._id}).then(()=>{
-						getSignal().then((res:any)=>{
-							initTable(res)
+				const swalWithBootstrapButtons = Swal.mixin({
+					customClass: {
+						confirmButton: 'btn btn-success',
+						cancelButton: 'btn btn-error'
+					},
+					buttonsStyling: false
+				})
 
-						})					})
+				swalWithBootstrapButtons.fire({
+					title: 'هل انت متأكد؟',
+					text: "هل تريد هل تريد حظر المستخدم ",
+					icon: 'warning',
+					showCancelButton: true,
+					confirmButtonText: 'نعم',
+					cancelButtonText: 'لا',
+					reverseButtons: true
+				}).then((result) => {
+					if (result.isConfirmed) {
+						blockUser.current=userId[data.index]
+						deleteProf({_id: blockUser.current}).then(() => setUserToHistory({
+							date: moment().format('MMMM Do YYYY, h:mm:ss a'),
+							adminID: userConnect.user._id,
+							userId: blockUser.current,
+							data: 'شكوى',
+							type: 'delete'
+						})).then(()=>{
+							deleteSignal({id:test[data.index]._id}).then(()=>{
+								getSignal().then((res:any)=>{
+									initTable(res)
+
+								})					})
+						})
+						Swal.fire({
+							position: 'center',
+							icon: 'success',
+							title: 'لقد تم الحظر',
+							showConfirmButton: false,
+							timer: 1000
+						})
+
+					} else if (
+						/* Read more about handling dismissals below */
+						result.dismiss === Swal.DismissReason.cancel
+					) {
+						Swal.fire({
+							position: 'center',
+							icon: 'error',
+							title: 'تم الغاء العملية',
+							showConfirmButton: false,
+							timer: 1000
+						})
+					}
 				})
